+++ project-frontend/src/components/Auth/Login.jsx
@@ -4,7 +4,7 @@ import axios from 'axios';
 import { ToastContainer } from 'react-toastify';
 import { toast } from 'react-toastify';
 import 'react-toastify/dist/ReactToastify.css';
-import backgroundImg from '../assets/background.jpg';
+import backgroundImg from '../../assets/background.jpg';
 
 function Login() {
   const navigate = useNavigate();
